+++ extensions/google-chrome/CHANGELOG.md
@@ -1,13 +1,18 @@
 # Google Chrome Changelog
 
+## [Search Bookmarks action] - 2022-12-17
+
+- Added search bookmarks action
+- Refactored code to use new useSQL hook
+
 ## [Open Tab feature] - 2021-11-08
 
-Add Open Tab action to action panel
+- Add Open Tab action to action panel
 
 ## [History] - 2021-10-26
 
-Add Google Chrome history search
+- Add Google Chrome history search
 
 ## [Initial Version] - 2021-10-20
 
-Add Google Chrome extension
+- Add Google Chrome extension